refactor(search): extract search component
